+++ src/com/apmath/template/domain/services/PostsService.kt
@@ -1,4 +1,8 @@
-package com.apmath.template.domain.posts
+package com.apmath.template.domain.services
+
+import com.apmath.template.domain.fetchers.PostsFetcherInterface
+import com.apmath.template.domain.posts.Post
+import com.apmath.template.domain.posts.PostInterface
 
 class PostsService(postFetcher: PostsFetcherInterface): PostsServiceInterface {
     private val postFetcher: PostsFetcherInterface = postFetcher
