--- src/domain/posts/PostsServiceInterface.kt
@@ -1,5 +1,0 @@
-package com.apmath.template.domain.posts
-
-interface PostsServiceInterface {
-    suspend fun get(id: Int): PostInterface
-}
